pyproject.toml
[tool.poetry]
name = "vertex-ai-template"
version = "0.1.0"
description = ""
license = "MIT"
authors = ["sfujiwara <shuhei.fujiwara@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <3.11"
kfp = "2.0.0b13"
functions-framework = "^3.3.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
jupyter = "^1.0.0"
google-cloud-aiplatform = "^1.15.1"
google-cloud-bigquery = "^2.34.4"
invoke = "^2.0.0"
toml = "^0.10.2"
tensorflow = "^2.10.0"
tensorflow-datasets = "^4.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
